BackEnd/requirements.txt
Django==4.1
djangorestframework==3.13.1
httpretty==1.1.4
mock==4.0.3
twilio==7.13.0